requirements.txt
langchain==0.3.5
langchain-community==0.3.3
langchain-core==0.3.13
langchain-openai==0.2.4
langchain-text-splitters==0.3.1
unstructured==0.16.3
gradio==5.5.0
langchain_chroma==0.1.4
sentence_transformers==3.2.1
python-dotenv==1.0.1
python-magic-bin==0.4.14
python-docx==1.1.2
chromadb~=0.5.18